Refs #21082. Refactor on `LocatorSelectorSender`.
