+++ src/cpp/rtps/writer/LocatorSelectorSender.cpp
@@ -1,5 +1,6 @@
 #include <fastdds/rtps/writer/LocatorSelectorSender.hpp>
-#include <fastdds/rtps/writer/RTPSWriter.hpp>
+
+#include <rtps/writer/BaseWriter.hpp>
 
 namespace eprosima {
 namespace fastdds {
