+++ test/unittest/statistics/rtps/RTPSStatisticsTests.cpp
@@ -45,6 +45,7 @@
 
 #include <rtps/participant/RTPSParticipantImpl.h>
 #include <rtps/transport/test_UDPv4Transport.h>
+#include <rtps/writer/BaseWriter.hpp>
 #include <statistics/rtps/monitor-service/Interfaces.hpp>
 #include <statistics/types/monitorservice_types.hpp>
 #include <statistics/types/types.hpp>
@@ -1408,7 +1409,8 @@ TEST_F(RTPSStatisticsTests, iconnections_queryable_get_entity_connections)
     for (auto& locator : conns_reader[0].announced_locators())
     {
         bool found = false;
-        for (auto& writer_loc : writer_->get_general_locator_selector().locator_selector)
+        auto base_writer = static_cast<fastdds::rtps::BaseWriter*>(writer_);
+        for (auto& writer_loc : base_writer->get_general_locator_selector().locator_selector)
         {
             //! Checking the address can be confusing since the writer_loc could be translated to
             //! 127.0.0.1
